+++ scripts/sct_propseg.py
@@ -21,12 +21,13 @@ import scipy
 
 import msct_image
 import msct_parser
+import sct_convert
 import sct_image
 import sct_label_utils
 import sct_utils as sct
 
 
-def check_and_correct(fname_segmentation, fname_centerline, threshold_distance=5.0, remove_temp_files=1, verbose=0):
+def check_and_correct_segmentation(fname_segmentation, fname_centerline, threshold_distance=5.0, remove_temp_files=1, verbose=0):
     """
     This function takes the outputs of isct_propseg (centerline and segmentation) and check if the centerline of the
     segmentation is coherent with the centerline provided by the isct_propseg, especially on the edges (related
@@ -39,11 +40,11 @@ def check_and_correct(fname_segmentation, fname_centerline, threshold_distance=5
 
     Returns: None
     """
-
+    sct.printv('\nCheck consistency of segmentation...', verbose)
     # creating a temporary folder in which all temporary files will be placed and deleted afterwards
     path_tmp = sct.tmp_create(verbose=verbose)
-    shutil.copy(fname_segmentation, path_tmp + 'tmp.segmentation.nii.gz')
-    shutil.copy(fname_centerline, os.path.join(path_tmp, 'tmp.centerline.nii.gz'))
+    sct_convert.convert(fname_segmentation, path_tmp + 'tmp.segmentation.nii.gz', squeeze_data=False, verbose=0)
+    sct_convert.convert(fname_centerline, path_tmp + 'tmp.centerline.nii.gz', squeeze_data=False, verbose=0)
 
     # go to tmp folder
     os.chdir(path_tmp)
@@ -74,13 +75,13 @@ def check_and_correct(fname_segmentation, fname_centerline, threshold_distance=5
 
     # for each slice of the segmentation, check if only one object is present. If not, remove the slice from segmentation.
     # If only one object (the spinal cord) is present in the slice, check if its center of mass is close to the centerline of isct_propseg.
+    slices_to_remove = [False] * nz  # flag that decides if the slice must be removed
     for i in range(nz):
         # extraction of slice
         slice = im_seg.data[:, :, i]
-        to_remove = False  # flag that decides if the slice must be removed
         label_objects, nb_labels = scipy.ndimage.label(slice)  # count binary objects in the slice
         if nb_labels > 1:  # if there is more that one object in the slice, the slice is removed from the segmentation
-            to_remove = True
+            slices_to_remove[i] = True
         elif nb_labels == 1:  # check if the centerline is coherent with the one from isct_propseg
             x_centerline, y_centerline = scipy.ndimage.measurements.center_of_mass(slice)
             slice_nearest_coord = min(key_centerline, key=lambda x:abs(x-i))
@@ -90,10 +91,29 @@ def check_and_correct(fname_segmentation, fname_centerline, threshold_distance=5
                                ((i - slice_nearest_coord) * pz) ** 2)
 
             if distance >= threshold_distance:  # threshold must be adjusted, default is 5 mm
-                to_remove = True
+                slices_to_remove[i] = True
 
+    # Check list of removal and keep one continuous centerline (improve this comment)
+    # Method:
+    # starting from mid-centerline (in both directions), the first True encountered is applied to all following slices
+    slice_to_change = False
+    for i in range(nz / 2, nz):
+        if slice_to_change:
+            slices_to_remove[i] = True
+        elif slices_to_remove[i]:
+            slices_to_remove[i] = True
+            slice_to_change = True
+    slice_to_change = False
+    for i in range(nz / 2, -1, -1):
+        if slice_to_change:
+            slices_to_remove[i] = True
+        elif slices_to_remove[i]:
+            slices_to_remove[i] = True
+            slice_to_change = True
+
+    for i in range(nz):
         # remove the slice
-        if to_remove:
+        if slices_to_remove[i]:
             im_seg.data[:, :, i] *= 0
 
     # saving the image
@@ -104,6 +124,8 @@ def check_and_correct(fname_segmentation, fname_centerline, threshold_distance=5
     sct.run('sct_image -i tmp.segmentation_RPI_c.nii.gz -setorient ' + image_input_orientation + ' -o ../' + fname_segmentation, verbose)
 
     os.chdir('..')
+
+    # display information about how much of the segmentation has been corrected
 
     # remove temporary files
     if remove_temp_files:
@@ -289,11 +311,11 @@ def main(args=None):
     parser = get_parser()
     arguments = parser.parse(args)
 
-    input_filename = arguments["-i"]
+    fname_data = arguments["-i"]
     contrast_type = arguments["-c"]
 
     # Building the command
-    isct_options = " -i " + input_filename + " -t " + contrast_type
+    isct_options = " -i " + fname_data + " -t " + contrast_type
 
     if "-ofolder" in arguments:
         folder_output = sct.slash_at_the_end(arguments["-ofolder"], slash=1)
@@ -376,23 +398,25 @@ def main(args=None):
     if "-alpha" in arguments:
         isct_options += " -alpha " + str(arguments["-alpha"])
 
-    # check if input image is in 3D. Otherwise itk image reader will cut the 4D image in 3D volumes and only take the
-    # first one.
-    image_input = msct_image.Image(input_filename)
+    image_input = msct_image.Image(fname_data)
     nx, ny, nz, nt, px, py, pz, pt = image_input.dim
     if nt > 1:
         sct.printv('ERROR: your input image needs to be 3D in order to be segmented.', 1, 'error')
 
-    path_fname, file_fname, ext_fname = sct.extract_fname(input_filename)
+    path_data, file_data, ext_data = sct.extract_fname(fname_data)
 
     # if centerline or mask is asked using viewer
     if use_viewer:
         # make sure image is in SAL orientation, as it is the orientation used by PropSeg
+
+
+
+
         image_input_orientation = sct_image.orientation(image_input, get=True, verbose=False)
-        path_fname, file_fname, ext_fname = sct.extract_fname(input_filename)
+        path_fname, file_fname, ext_fname = sct.extract_fname(fname_data)
         reoriented_image_filename = 'tmp.' + sct.add_suffix(file_fname + ext_fname, "_SAL")
         path_tmp_viewer = sct.tmp_create(verbose=verbose)
-        sct_image.main(['-i', input_filename,
+        sct_image.main(['-i', fname_data,
                         '-o', os.path.join(path_tmp_viewer, reoriented_image_filename),
                         '-setorient', 'SAL',
                         '-v', '0'])
@@ -423,7 +447,7 @@ def main(args=None):
             # reorient the initialization mask to correspond to input image orientation
             mask_reoriented_filename = sct.add_suffix(file_fname + ext_fname, "_mask_viewer")
             sct_image.main(['-i', os.path.join(path_tmp_viewer, mask_filename),
-                            '-o', os.path.join(path_tmp_viewer, mask_filename),
+                            '-o', os.path.join(folder_output, mask_reoriented_filename),
                             '-setorient ' + image_input_orientation,
                             '-v', '0'])
 
@@ -433,22 +457,33 @@ def main(args=None):
             elif use_viewer == "mask":
                 isct_options += " -init-mask " + folder_output + mask_reoriented_filename
         else:
-            sct.printv('\nERROR: the viewer has been closed before entering all manual points. Please try again.',
-                       verbose, type='error')
+            sct.printv('\nERROR: the viewer has been closed before entering all manual points. Please try again.', 1,
+                       type='error')
 
-    cmd = 'isct_propseg  {0} -centerline-binary'.format(isct_options)
-    sct.run(cmd, verbose)
+    isct_options += ' -centerline-binary'
+    cmd = 'isct_propseg' + isct_options
+    status, output = sct.run(cmd, verbose, error_exit='verbose')
 
-    # extracting output filename
-    path_fname, file_fname, ext_fname = sct.extract_fname(input_filename)
-    output_filename = file_fname + "_seg" + ext_fname
-    fname_centerline = file_fname + '_centerline' + ext_fname
-    print 33*'*',folder_output + output_filename
-    print folder_output + fname_centerline
-    print remove_temp_files
-    print use_viewer
-    print 33*'*'
-    check_and_correct(folder_output + output_filename, folder_output + fname_centerline, remove_temp_files)
+    # check status is not 0
+    if not status == 0:
+        sct.printv('\nERROR: Automatic cord detection failed. Please initialize using -init-centerline or -init-mask (see help).', 1, type='error')
+
+    # build output filename
+    file_seg = file_data + "_seg" + ext_data
+    if folder_output == "./":
+        fname_seg = file_seg
+    else:
+        fname_seg = folder_output + file_seg
+
+    # check consistency of segmentation
+    fname_centerline = folder_output + file_data + '_centerline' + ext_data
+    check_and_correct_segmentation(fname_seg, fname_centerline, threshold_distance=3.0, remove_temp_files=remove_temp_files, verbose=verbose)
+
+    # copy header from input to segmentation to make sure qform is the same
+    from sct_image import copy_header
+    im_seg = msct_image.Image(fname_seg)
+    im_seg = copy_header(image_input, im_seg)
+    im_seg.save(data_type='int8')
 
     # remove temporary files
     if remove_temp_files:
@@ -456,11 +491,8 @@ def main(args=None):
         if use_viewer:
             shutil.rmtree(path_tmp_viewer, ignore_errors=True)
 
-    if folder_output == "./":
-        output_name = output_filename
-    else:
-        output_name = folder_output + output_filename
-    sct.printv("fslview " + input_filename + " " + output_name + " -l Red -b 0,1 -t 0.7 &\n", verbose, 'info')
+    sct.printv('\nDone! To view results, type:', verbose)
+    sct.printv("fslview "+fname_data+" "+fname_seg+" -l Red -b 0,1 -t 0.7 &\n", verbose, 'info')
 
 
 if __name__ == "__main__":