+++ scripts/sct_register_graymatter.py
@@ -151,7 +151,8 @@ class MultiLabelRegistration(object):
         fname_warp_multilabel_template2auto = 'warp_' + file_template_ml + '2' + file_automatic_ml + '.nii.gz'
         fname_warp_multilabel_auto2template = 'warp_' + file_automatic_ml + '2' + file_template_ml + '.nii.gz'
 
-        self.fname_warp_template2gm = 'warp_template2' + file_gm + '.nii.gz'
+        self.fname_warp_template2gm = sct.extract_fname(self.fname_warp_template2target)[1] + '_reg_gm' +\
+                                      sct.extract_fname(self.fname_warp_template2target)[2]
 
         sct.run('sct_concat_transfo -w ' + file_warp_template2target + ext_warp_template2target + ',' +
                 fname_warp_multilabel_template2auto + ' -d ' + file_gm + ext_gm + ' -o ' + self.fname_warp_template2gm)
@@ -164,9 +165,10 @@ class MultiLabelRegistration(object):
             elif self.template == 'PAM50':
                 fname_dest = path_sct + '/data/PAM50/template/PAM50_t2.nii.gz'
 
-            self.fname_warp_gm2template = 'warp_' + file_gm + '_gm2template.nii.gz'
-            sct.run('sct_concat_transfo -w ' + fname_warp_multilabel_auto2template + ',' + file_warp_target2template +
-                    ext_warp_target2template + ' -d ' + fname_dest + ' -o ' + self.fname_warp_gm2template)
+            self.fname_warp_gm2template = sct.extract_fname(self.fname_warp_target2template)[1] \
+                                          + '_reg_gm' +sct.extract_fname(self.fname_warp_target2template)[2]
+            sct.run('sct_concat_transfo -w '+fname_warp_multilabel_auto2template+','+file_warp_target2template +
+                    ext_warp_target2template+' -d '+fname_dest+' -o '+self.fname_warp_gm2template)
 
         os.chdir('..')
 
@@ -472,40 +474,34 @@ def visualize_warp(fname_warp, fname_grid=None, step=3, rm_tmp=True):
 def get_parser():
     # Initialize the parser
     parser = msct_parser.Parser(__file__)
-    parser.usage.set_description('Multi-label registration\n')
-    parser.add_option(
-        name="-gm",
-        type_value="file",
-        description="Gray matter automatic segmentation",
-        mandatory=True,
-        example='t2star_gmseg.nii.gz')
-    parser.add_option(
-        name="-wm",
-        type_value="file",
-        description="White matter automatic segmentation",
-        mandatory=True,
-        example='t2star_wmseg.nii.gz')
-    parser.add_option(
-        name="-t",
-        type_value="folder",
-        description="Path to template (registered on target image)",
-        mandatory=False,
-        default_value='label/')
-    parser.add_option(
-        name="-w",
-        type_value="file",
-        description="Warping field template --> target image",
-        mandatory=True,
-        example='warp_template2t2star.nii.gz')
+    parser.usage.set_description('Registration function to improve the template registration by accounting for the gray and white matter shape using a multi-label approach. Output is a warping field from the template to the target image accounting for the gray matter shape. If -winv is used, output also includes the inverse warping field (from the target image to the template) that accounts for the gray matter shape.')
+    parser.add_option(name="-gm",
+                      type_value="file",
+                      description="Gray matter automatic segmentation",
+                      mandatory=True,
+                      example='t2star_gmseg.nii.gz')
+    parser.add_option(name="-wm",
+                      type_value="file",
+                      description="White matter automatic segmentation",
+                      mandatory=True,
+                      example='t2star_wmseg.nii.gz')
+    parser.add_option(name="-t",
+                      type_value="folder",
+                      description="Path to template (registered on target image)",
+                      mandatory=False,
+                      default_value='label/')
+    parser.add_option(name="-w",
+                      type_value="file",
+                      description="Warping field template --> target image",
+                      mandatory=True,
+                      example='warp_template2t2star.nii.gz')
 
-    parser.add_option(
-        name="-param",
-        type_value="str",
-        description="Parameters for the multimodal registration between multilabel images",
-        mandatory=False,
-        default_value=Param().param_reg,
-        example='step=1,algo=slicereg,metric=MeanSquares,step=2,algo=syn,metric=MeanSquares,iter=2:step=3,algo=bsplinesyn,metric=MeanSquares,iter=5,smooth=1'
-    )
+    parser.add_option(name="-param",
+                      type_value="str",
+                      description="Parameters for the multimodal registration between multilabel images",
+                      mandatory=False,
+                      default_value=Param().param_reg,
+                      example='step=1,algo=slicereg,metric=MeanSquares,step=2,algo=syn,metric=MeanSquares,iter=2:step=3,algo=bsplinesyn,metric=MeanSquares,iter=5,smooth=1')
 
     parser.usage.addSection('\nOUTPUT OTIONS')
     parser.add_option(