+++ CHANGES.md
@@ -1,6 +1,8 @@
 # CHANGES TO RELEASE
 
-##3.0_beta30 (2017-01-08)
+##3.0_beta31 (2017-01-16)
+- BUG: **sct_process_segmentation** display spinal cord length when required (full spinal cord) (#1112)
+- BUG: **sct_propseg** Adding rules for in-segmentation errors (#1110)
 - BUG: PAM50: probabilist templates of WM and GM not sum to 1 (#411)
 - BUG: **sct_propseg**: fixed edge issue (#1074)
 - BUG: **sct_label_vertebrae**: fixed temporary folder not created (#1083)
@@ -9,7 +11,7 @@
 - BUG: **sct_register_graymatter**: fixed empty inverse warping field (#1068)
 - OPT: **sct_label_vertebrae**: now outputing verbose=2 pics in ofolder (#1094)
 - OPT: **sct_straighten_spinalcord**: fixed Reduce required RAM (#979)
-- OPT: **sct_straighten_spinalcord**: removes accuracy computation by default (#1075)
+- OPT: **sct_straighten_spinalcord**: removes accuracy computation by default (#1075) 
 - OPT: **sct_propseg**: improve robustness towards exception (#207)
 - OPT: **isct_test_function**: send email when finished (#1081)
 - OPT: **isct_test_function**: removed color tags on log entries  (#1035)